Version 2.0 Rock, Paper, Scissors
The previous code would not run correctly - research how to create a "while" loop in order to make the programme start again. This version works but has some obvious issues - does not display an invalid code when the user inputs a spelling error when writing "Rock, Paper or Scissors". Also, does not keep rounds - might add this in V3.